+++ pkgs/cli_config/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1
 
 - Support non-String map keys in YAML configuration files.
+- Support null values in YAML configuration files.
 
 ## 0.1.0